[ft] add secret word feature to be used on password reset
The secret word will be hashed on entry into the db and will only be known
to the user of the account. For a user to be able to reset their passwords,
they will be prompted to provide their secret keys, their emails and their new
passwords if the secret key does not mach or the email does not match, an error
is thrown.
The commit also incorporates the secret word in the test files